Criação da controller para Nota

+++ src/Models/Nota.cs
@@ -5,8 +5,8 @@ namespace ControleNotas.src.Models
         public int Id { get; set; }
         public int AlunoId { get; set; }
         public int DisciplinaId { get; set; }
-        public int Nota1 { get; set; }
-        public int Nota2 { get; set; }
+        public double Nota1 { get; set; }
+        public double Nota2 { get; set; }
 
         public Aluno Aluno { get; set; } = null!;
         public Disciplina Disciplina { get; set; } = null!;
